requirements.txt
anthropic>=0.39.0
flask>=3.0.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.8.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
